requirements.txt
pyautogui
keyboard
pytesseract
Pillow
